+++ Sources/Lux/InjectorImplementations/Zsh/ZshInjector.swift
@@ -8,7 +8,7 @@ extension RegexPattern {
           #""[^"]*"|'[^']*'"# // strings
         + #"|\s?\x5C#[^\s]*|#.*(?=\n|\Z)"# // comments and escaped # signs
         + #"|(\s?sudo|\$\(|\$|\[|`|\n|\r|\|)\h*[a-zA-Z0-9]{1}[a-zA-Z0-9_-]*=?"# // programs and variables defs
-        + #"|\s\h*[a-zA-Z0-9\/\.]{1}[^\s]*"# // commands and option values
+        + #"|\s\h*[a-zA-Z0-9\/\.]{1}[^\s^`^\(^\)]*"# // commands and option values
         + #"|\s-{1,2}[a-zA-Z0-9_-]+"# // options and flags
         + #"|\$\{[a-zA-Z0-9_-]+\}|"# // variable with brackets ${variable}
         + #"\[|\]|;|\(|\)|\{|\}|`"#, // punctuation