+++ Sources/Lux/Themes/Dracula/DraculaThemeInjectorDelegate.swift
@@ -1,0 +1,5 @@
+protocol DraculaThemeInjectorDelegate: ThemeInjectorDelegate {}
+
+extension DraculaThemeInjectorDelegate {
+    var themeBackgroundColor: Color { Color.dracula.background }
+}